Add new version of useGamma hooks and DigitGammaActions
These hooks rely on the backend implementing sessions via cookies, and them having two api endpoints. One GET /me, to give information about the signed in user. This is the check to see if the user is actually signed in. If failed, then the url to gamma with correct client id and so on should be sent as a response along with a 403 error code. Also one POST /auth that the frontend sends the code that is received after being redirected from a successful login at gamma.